git-svn-id: svn://localhost/ardour2/branches/undo@794 d708f5d6-7413-0410-9779-e7cbd77b26cf

+++ libs/ardour/track.cc
@@ -152,7 +152,7 @@ bool
 Track::can_record()
 {
 	bool will_record = true;
-	for (int i = 0; i < _inputs.size() && will_record; i++) {
+	for (size_t i = 0; i < _inputs.size() && will_record; i++) {
 		if (!_inputs[i]->connected())
 			will_record = false;
 	}
